# use vImage framework to accelerate conversion from YUV to RGB(A)