Hi. I am working at Link-U, a company which publishes digital comics in Japan.

In this pull request, I added an optimized code path to speed up conversions from YUV to RGB(A) using [vImage framework](https://developer.apple.com/documentation/accelerate/conversion?language=objc).

Unfortunately, it seems that vImage conversion API does not provide fully support for 10bit/12bit YUV image. Hence, in this patch, conversions from only 8bit YUV images are accelerated.

## Benchmark result

All results are measured in iPhone 6, iOS 12.2.

### Original version

[Here](https://github.com/link-u/SDWebImageAVIFCoder/tree/wip/measure-original) is a branch with measuring code.

```
# 8-bit image
SDWebImageAVIFCoder_Example[765:156941] Converting from YUV to RGB takes 0.174398 sec
SDWebImageAVIFCoder_Example[765:156924] Static AVIF load success
# 10-bit image
SDWebImageAVIFCoder_Example[765:156941] Converting from YUV to RGB takes 0.465885 sec
SDWebImageAVIFCoder_Example[765:156924] HDR AVIF load success
```

### vImage version(this patch)

[Here](https://github.com/link-u/SDWebImageAVIFCoder/tree/wip/measure-vImage) is a branch with measuring code.

```
# 8-bit image
SDWebImageAVIFCoder_Example[768:157661] Converting from YUV to RGB takes 0.007166 sec
SDWebImageAVIFCoder_Example[768:157548] Static AVIF load success
# 10-bit image
SDWebImageAVIFCoder_Example[768:157661] Converting from YUV to RGB takes 0.460658 sec
SDWebImageAVIFCoder_Example[768:157548] HDR AVIF load success
 ```

This patch makes a conversion more than 24 times faster, in 8-bit image.

Please take a look!