@ledyba-z 2020-01-19
P.S.

Monochrome images are also considered in this patch, however, libavif does not handle them correctly.

I sent a pull request about this issue to libavif:
https://github.com/AOMediaCodec/libavif/pull/39