@dreampiggy 2020-01-19
Seems good...
Previouslly, to support HDR or 12bit AVIF image (I test it using the https://github.com/AOMediaCodec/av1-avif/tree/master/testFiles), I have to use the CPU code to do this :)

If the vImage can have a better performance, I'd like to replace that.