@ledyba-z 2020-01-19
Thank you!

For example, there is no conversion API from 10bit/12bit YUV420 to RGB888.
However, there can be some alternative way to convert (for example, converting 10bit YUV420 → 10bit YUV444 → 8bit RGB or 16bit RGB).
I will try them.

By the way, what should 10-bit/12-bit images be converted to?
Currently, 10bit/12bit images also seem to be converted to 8bit RGB(A):
https://github.com/SDWebImage/SDWebImageAVIFCoder/blob/ca147aac078ccdf391ab93c3133cea07f57c6e64/SDWebImageAVIFCoder/Classes/SDImageAVIFCoder.m#L164
How about converting them to 16bit RGB(A)?