@ledyba-z 2020-01-21
@dreampiggy 
Hi, I succeeded to support 10bit/12bit images.

```
benchmark @ iPhone 6

# before
SDWebImageAVIFCoder_Example[768:157661] Converting from YUV to RGB takes 0.460658 sec
SDWebImageAVIFCoder_Example[768:157548] HDR AVIF load success
# after
SDWebImageAVIFCoder_Example[713:130564] Converting from YUV to RGB takes 0.090242 sec
SDWebImageAVIFCoder_Example[713:130433] HDR AVIF load success
```

10bit/12bit versions seems strongly optimized than 8bit versions yet, but it makes about 5 times faster!

By the way, test images in some format are not available at [official repo](https://github.com/AOMediaCodec/av1-avif/tree/master/testFiles):

 - 8bit/10bit/12bit YUV422
 - 10bit/12bit monochrome

So, I prepared some examples images by myself using [cavif](https://github.com/link-u/cavif), so please use them to check:

https://github.com/link-u/avif-sample-images