@dreampiggy 2020-01-21
I’ll test it soon. Sorry for your patient.

I’m in China and on vacation in China’s lunar new year. If I have a time to check, I’ll merge and release a new version,
.