@dreampiggy 2020-01-22
Please merge the master branch again. I fixed the Travis-CI issue and support SwiftPM now. You can check whether this dev branch works on different compile case. (For example, SwiftPM/Carthage does not have dav1d or rav1e, always use libavif method and don’t directly use them)