@dreampiggy 2020-01-22
@ledyba-z libavif released new v0.5.4, does that fix the issue ?


https://github.com/AOMediaCodec/libavif/commit/e080130e1c9975d1f3dac0c159ed8000ff5ca244