@ledyba-z 2020-01-23
> @ledyba-z libavif released new v0.5.4, does that fix the monochrome issue ?
> https://github.com/AOMediaCodec/libavif/blob/master/CHANGELOG.md
>  AOMediaCodec/libavif@e080130

Yeah, applied patch was the same I sent, so the monochrome issue shall be fixed.