@dreampiggy 2020-01-24
@ledyba-z Hi. The 0.5.0 version is released, check: [v0.5.0 changelog](https://github.com/SDWebImage/SDWebImageAVIFCoder/releases/tag/0.5.0) for details.

Thanks for your support, you're actually the first contributor here, I've updated the Readme with this :)