@dreampiggy 2020-01-24
I'm decided to support thumbnail decoding soon in v0.6.0 (mentioned in [SDWebImage 5.5.0 new feature](https://github.com/SDWebImage/SDWebImage/releases/tag/5.5.0), which need each codec update to support). See the similiar update to [WebPCoder](https://github.com/SDWebImage/SDWebImageWebPCoder/pull/31)

If you have some better idea about this, please contact me, thanks.