@dreampiggy 2020-01-24
One possible issue I don't actually realized. The libavif 0.5.0 provide the API to grab the EXIF/XMP metadata. Does AVIF support EXIF orientation like JPEG ?

If so, maybe we should update the code to provide a fix for the non-Up AVIF images.