@ledyba-z 2020-01-26
@dreampiggy Thank you, you are welcome!

> support thumbnail decoding for AVIF

[MIAF](https://www.iso.org/standard/74417.html), a file format for AVIF and HEIF, can hold thumbnail images (it's called `MIAF thumbnail images`).

However, there seems not to be any example files with thumbnails yet, and libavif might not support decoding thumbnail images yet. 

> Does AVIF support EXIF orientation like JPEG ?

YES, but it's not need to parse EXIF.

AVIF supports orientation/mirroing/crop itself. And also, MIAF specifies that readers are not needed to parse exif metadata to rotate/mirror/crop (If you want to read other metadata, you still have to read exif).

I prepared the test files, so please check [our example files](https://github.com/link-u/avif-sample-images#kimono---transformation-tests).

Our encoder, [cavif](https://github.com/link-u/cavif), and decoder, [davif](https://github.com/link-u/davif), supports those features. But libavif looks not support them at all. I will report an issue.